Refactoring lbs_structs.h
