+++ modules/linear_boltzmann_solvers/lbs_problem/iterative_methods/wgs_context.h
@@ -3,8 +3,8 @@
 
 #pragma once
 
-#include "modules/linear_boltzmann_solvers/lbs_problem/lbs_structs.h"
 #include "framework/math/linear_solver/linear_solver_context.h"
+#include "modules/linear_boltzmann_solvers/lbs_problem/source_functions/source_flags.h"
 #include <vector>
 #include <functional>
 #include <memory>
